cleanup issue handling
