+++ src/lib/autoinstall/autoinst_issues/ay_missing_value.rb
@@ -17,25 +17,27 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
-require "autoinstall/autoinst_issues/issue"
+require "installation/autoinst_issues/issue"
 
-module Y2Autoinstallation
+module Installation
   module AutoinstIssues
-    # Represents an AutoYaST situation where an invalid value was given.
+    # Represents an AutoYaST situation where a mandatory value is missing.
     #
-    class InvalidValue < Issue
-      include Yast::Logger
+    # @example Missing value for attribute 'bar' in 'foo' section.
+    #   problem = AyMissingValue.new("foo","bar")
+    class AyMissingValue < ::Installation::AutoinstIssues::Issue
+      attr_reader :section, :attribute
+      attr_reader :description, :severity
 
-      # @param section     [String] main section name in the AutoYaST configuration file
-      # @param attribute   [String] wrong attribute
-      # @param value       [String] wrong attribute value
-      # @param description [String] additional explanation
-      # @param severity    [Symbol] :warn, :fatal = abort the installation
-      def initialize(section, attribute, value, description, severity = :warn)
+      # @param section     [String] Section where it was detected
+      # @param attribute   [String] Name of the missing attribute
+      # @param description [String] additional explanation; optional
+      # @param severity    [Symbol] :warn, :fatal = abort the installation ; optional
+      def initialize(section, attr, description = "", severity = :warn)
         textdomain "autoinst"
+
         @section = section
-        @attribute = attribute
-        @value = value
+        @attribute = attr
         @description = description
         @severity = severity
       end
@@ -46,10 +48,9 @@ module Y2Autoinstallation
       # @see Issue#message
       def message
         # TRANSLATORS:
-        # 'value' is a generic value (number or string) 'attribute' is an AutoYaST element
+        # 'attr' is an AutoYaST element
         # 'description' has already been translated in other modules.
-        format(_("Invalid value '%{value}' for attribute '%{attribute}': %{description}"),
-          value: @value, attribute: @attribute, description: @description)
+        format(_("Missing element '%{attr}'. %{description}"), attr: attr, description: description)
       end
     end
   end
